+++ webapp/src/app/shared/phaser/entities/cc-entity.ts
@@ -1,12 +1,12 @@
-import {CCMap} from '../tilemap/cc-map';
-import {MapEntity, Point, Point3} from '../../../models/cross-code-map';
-import {Helper} from '../helper';
+import { CCMap } from '../tilemap/cc-map';
+import { MapEntity, Point, Point3 } from '../../../models/cross-code-map';
+import { Helper } from '../helper';
 import * as Phaser from 'phaser';
-import {Vec2} from '../vec2';
+import { Vec2 } from '../vec2';
 
-import {Globals} from '../../globals';
-import {BaseObject} from '../base-object';
-import {Subscription} from 'rxjs';
+import { Globals } from '../../globals';
+import { BaseObject } from '../base-object';
+import { Subscription } from 'rxjs';
 
 export interface ScaleSettings {
 	scalableX: boolean;
@@ -28,6 +28,12 @@ export interface AttributeValue {
 	optional?: boolean;
 	
 	// TODO: not needed anymore, cc source is not obfuscated
+	[key: string]: any;
+}
+
+export interface DetailSettings {
+	mapId: number;
+	
 	[key: string]: any;
 }
 
@@ -72,13 +78,19 @@ export abstract class CCEntity extends BaseObject {
 	get selected(): boolean {
 		return this._selected;
 	}
-
+	
 	// drag
 	public isDragged = false;
 	public startOffset: Point = {x: 0, y: 0};
 	
 	// zIndex: number;
-	details: { level: { level: number, offset: number }, type: string, settings: any } = <any>{};
+	details: {
+		level: {
+			level: number;
+			offset: number;
+		}; type: string;
+		settings: DetailSettings;
+	} = <any>{};
 	entitySettings: {
 		collType: string;
 		baseSize: Point3;
@@ -277,8 +289,8 @@ export abstract class CCEntity extends BaseObject {
 					}
 					
 					// crop offset
-					img.x -= cropX;
-					img.y -= cropY;
+					img.x -= sheet.x;
+					img.y -= sheet.y;
 					
 					img.setCrop(cropX, cropY, sheet.w, sheet.h);
 					img.flipX = !!sheet.flipX;